fix: レイアウトシフトが起こらないように修正
