+++ src/components/model/PrefectureCheckBoxes/presentations/css/base.css.ts
@@ -1,0 +1,8 @@
+import { style } from '@vanilla-extract/css';
+
+export const container = style({
+  display: 'flex',
+  flexWrap: 'wrap',
+  alignItems: 'center',
+  position: 'relative',
+});
